Default to internal_embedded_datasource if no cayenne.test.connection property is specified.  [CAY-274]
This is an embedded HSQLDB instance.   The classpath is modified appropriately to add hsqldb.jar in this case.

I'm not entirely thrilled about using "internal_embedded_datasource" as it could potentially conflict with a user name, but the build script needs to have a name so it can create a report directory.  (Originally I tried keeping cayenne.test.connection undefined)

I'm also not sure if CayenneTestResource is the best place to hardcode the DataSourceInfo.


git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/cayenne/main/trunk@403628 13f79535-47bb-0310-9956-ffa450edef68